requirements.txt
PyQt6>=6.6.0
openai>=1.0.0
anthropic>=0.39.0
keyring>=24.0.0
requests>=2.31.0
pyspellchecker>=0.8.0
python-docx>=1.1.0
